fix: make internal types private
